pyproject.toml
[tool.poetry]
name = "wmtscheduler-sync"
version = "0.1.0"
description = ""
authors = ["Carlton Yeung <CarltonYeung@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-api-python-client = "^2.70.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.8.0"
pdftotext = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
